+++ packages/styles/src/packages/atomic/iconButton/index.ts
@@ -1,0 +1,6 @@
+import { styles } from "./rarui-iconButton.css";
+
+export const iconButton = {
+  classnames: { ...styles },
+};
+export type { IconButtonVariants } from "./rarui-iconButton.types";