+++ packages/styles/src/packages/atomic/iconButton/rarui-iconButton.types.ts
@@ -1,0 +1,4 @@
+import { RecipeVariants } from "@vanilla-extract/recipes";
+import { styles } from "./rarui-iconButton.css";
+
+export type IconButtonVariants = RecipeVariants<typeof styles.iconButton>;